backend/requirements.txt
django>=4.2,<5.0
djangorestframework>=3.14,<4.0
gunicorn>=21.2  # wsgi server for production
django-cors-headers>=4.0  # to handle cross-origin queries
psycopg2-binary>=2.9  # for postgres
whitenoise>=6.5  # static files management in production

# live chat
channels>=4.0
channels-redis
daphne>=4.0

# 2FA et JWT
djangorestframework-simplejwt==5.3.1
pyotp==2.8.0
